feat: AppBar da pagina calendar
