+++ dart/lib/Activity/activity.dart
@@ -1,0 +1,43 @@
+import 'package:flutter/material.dart';
+
+class Activity extends StatefulWidget {
+  const Activity({super.key});
+
+  @override
+  State<Activity> createState() => _ActivityState();
+}
+
+class _ActivityState extends State<Activity> {
+  bool _favorited = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Theme.of(context).colorScheme.inversePrimary,
+      child: Column(children: [
+        Text(
+          'Activity title',
+          style: Theme.of(context).textTheme.bodySmall,
+        ),
+        const Text('A Física dos Buracos Negros Supermassivos'),
+        const Text('Mesa redonda'),
+        const Text('Domingo 07:00h - 08:00h'),
+        const Text('Sthepen William Hawking'),
+        const Text('Maputo'),
+        const Text('Astrofísica e Cosmologia'),
+        ElevatedButton.icon(
+          onPressed: () {
+            setState(() {
+              _favorited = !_favorited;
+            });
+          },
+          icon: _favorited
+              ? const Icon(Icons.star)
+              : const Icon(Icons.star_outline),
+          label: Text(
+              _favorited ? 'Remover da sua agenda' : 'Adicionar à sua agenda'),
+        )
+      ]),
+    );
+  }
+}
