+++ dart/pubspec.yaml
@@ -35,6 +35,11 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  dio: ^5.4.3+1
+  cached_network_image: ^3.3.1
+  from_css_color: ^2.0.0
+  go_router: ^13.2.4
+  json_annotation: ^4.8.1
 
 dev_dependencies:
   flutter_test:
